# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/sync/push": {
      "post": {
        "description": "Accepts a batch of operations from the client offline queue. Processes operations transactionally with idempotency support. Duplicate batches (same idempotencyKey) return cached response.",
        "operationId": "SyncController_push_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sync push processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Push offline changes to server",
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/v1/sync/pull": {
      "get": {
        "description": "Returns entities that have changed since the provided timestamp. Use the returned serverTimestamp as the since parameter for subsequent pulls. Includes surveys, sections, answers, and media metadata owned by the user.",
        "operationId": "SyncController_pull_v1",
        "parameters": [
          {
            "name": "since",
            "required": false,
            "in": "query",
            "description": "ISO 8601 timestamp to fetch changes after (exclusive)",
            "schema": {
              "example": "2024-01-15T10:30:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of changes to return (1-500, default 100)",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Changes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Pull server changes",
        "tags": [
          "Sync"
        ]
      }
    }
  },
  "info": {
    "title": "SurveyScriber API",
    "description": "SurveyScriber Backend API - Professional Field Survey & Inspection Platform\n\n## Authentication\nThis API uses JWT Bearer tokens. Obtain tokens via `/auth/login` and include in the Authorization header.\n\n## Offline-First Architecture\nThe API supports granular CRUD operations for surveys, sections, and answers to enable offline-first Flutter clients to sync individual changes without full survey uploads.",
    "version": "1.0.0",
    "contact": {
      "name": "SurveyScriber Team",
      "url": "https://surveyscriber.com",
      "email": "support@surveyscriber.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token from /auth/login response",
        "in": "header"
      }
    },
    "schemas": {
...
      "SyncOperationDto": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Client-generated unique ID for this operation (for idempotency)",
            "example": "op-123e4567-e89b-12d3-a456-426614174000"
          },
          "operationType": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE"
            ],
            "description": "Type of operation",
            "example": "CREATE"
          },
          "entityType": {
            "type": "string",
            "enum": [
              "SURVEY",
              "SECTION",
              "ANSWER",
              "MEDIA"
            ],
            "description": "Type of entity being synced",
            "example": "SURVEY"
          },
          "entityId": {
            "type": "string",
            "description": "UUID of the entity",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "data": {
            "type": "object",
            "description": "Entity data payload (for CREATE/UPDATE)",
            "example": {
              "title": "New Survey",
              "propertyAddress": "123 Main St"
            }
          },
          "clientTimestamp": {
            "type": "string",
            "description": "ISO timestamp when operation was created on client",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "operationId",
          "operationType",
          "entityType",
          "entityId"
        ]
      },
      "SyncPushDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "description": "Client-generated idempotency key for the entire batch",
            "example": "batch-123e4567-e89b-12d3-a456-426614174000"
          },
          "operations": {
            "description": "Array of sync operations to apply",
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncOperationDto"
            }
          }
        },
        "required": [
          "idempotencyKey",
          "operations"
        ]
      },
      "SyncOperationResultDto": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Client operation ID",
            "example": "op-123e4567-e89b-12d3-a456-426614174000"
          },
          "success": {
            "type": "boolean",
            "description": "Whether operation succeeded",
            "example": true
          },
          "error": {
            "type": "string",
            "description": "Error message if operation failed",
            "example": "Entity not found"
          },
          "entityId": {
            "type": "string",
            "description": "Server-side entity ID (may differ from client for new entities)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "duplicate": {
            "type": "boolean",
            "description": "Whether this operation was a duplicate (idempotent replay)",
            "example": false
          }
        },
        "required": [
          "operationId",
          "success"
        ]
      },
      "SyncPushResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Overall success status",
            "example": true
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Idempotency key echoed back",
            "example": "batch-123e4567-e89b-12d3-a456-426614174000"
          },
          "results": {
            "description": "Results for each operation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncOperationResultDto"
            }
          },
          "serverTimestamp": {
            "format": "date-time",
            "type": "string",
            "description": "Server timestamp when sync was processed",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "batchDuplicate": {
            "type": "boolean",
            "description": "Whether entire batch was replayed (duplicate idempotency key)",
            "example": false
          }
        },
        "required": [
          "success",
          "idempotencyKey",
          "results",
          "serverTimestamp"
        ]
      },
      "SyncEntityDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "enum": [
              "SURVEY",
              "SECTION",
              "ANSWER",
              "MEDIA"
            ],
            "description": "Type of entity",
            "example": "SURVEY"
          },
          "entityId": {
            "type": "string",
            "description": "Entity UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "changeType": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE"
            ],
            "description": "Type of change",
            "example": "UPDATE"
          },
          "data": {
            "type": "object",
            "description": "Entity data (null for DELETE)"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When entity was last updated on server",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "entityType",
          "entityId",
          "changeType",
          "data",
          "updatedAt"
        ]
      },
      "SyncPullResponseDto": {
        "type": "object",
        "properties": {
          "changes": {
            "description": "Array of changed entities",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncEntityDto"
            }
          },
          "serverTimestamp": {
            "format": "date-time",
            "type": "string",
            "description": "Server timestamp for this response (use as since for next pull)",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether there are more changes to fetch",
            "example": false
          },
          "totalCount": {
            "type": "number",
            "description": "Total count of changes matching query",
            "example": 15
          }
        },
        "required": [
          "changes",
          "serverTimestamp",
          "hasMore",
          "totalCount"
        ]
      }
    }
  }
}